travel_itinerary_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
mssql-django==1.4
pyodbc==5.0.1
python-decouple==3.8